requirements.txt
aniso8601==9.0.1
blinker==1.5
click==8.1.3
dnspython==2.3.0
email-validator==1.3.1
Flask==2.2.3
Flask-Login==0.6.2
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.1.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
passlib==1.7.4
pytz==2023.3.post1
six==1.16.0
SQLAlchemy==2.0.6
typing-extensions==4.5.0
Werkzeug==2.2.3
WTForms==3.0.1
zipp==3.15.0
